+++ packages/vendure-plugin-qls-fulfillment/CHANGELOG.md
@@ -1,3 +1,9 @@
-# 1.0.0 (2025-11-07) ## CHANGE THIS DATE AND VERSION
+# 1.1.0 (2026-01-07)
+
+- Fix delivery options to be an array of objects with a tag property
+- Emit event for failed order push
+- Delay 10s before setting custom field `syncedToQls` to prevent race conditions.
+
+# 1.0.0 (2025-11-07)
 
 - Initial release